Fix to convert `shf_mod64` to u64 when calling assign.

+++ zkevm-circuits/src/evm_circuit/util/math_gadget.rs
@@ -1142,7 +1142,7 @@ impl<F: Field> ShrWordsGadget<F> {
         self.shf_div64
             .assign(region, offset, Some(F::from(shf_div64 as u64)))?;
         self.shf_mod64
-            .assign(region, offset, Some(F::from(shf_mod64)))?;
+            .assign(region, offset, Some(F::from(shf_mod64 as u64)))?;
         self.p_lo.assign(region, offset, Some(F::from_u128(p_lo)))?;
         self.p_hi.assign(region, offset, Some(F::from_u128(p_hi)))?;
         self.shf_lt256
